@@ -3,36 +3,40 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.4.1",
-    "@mui/material": "^5.4.1",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "chart.js": "^4.3.0",
-    "highcharts": "^11.0.1",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.14.3",
+    "@mui/material": "^5.14.3",
+    "@mui/x-date-pickers": "^6.10.2",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "chart.js": "^4.3.3",
+    "date-fns": "^2.30.0",
+    "dayjs": "^1.11.9",
+    "highcharts": "^11.1.0",
     "highcharts-react-official": "^3.2.0",
-    "html-to-react": "^1.5.0",
+    "html-to-react": "^1.6.0",
     "html2canvas": "^1.4.1",
     "html2pdf.js": "^0.10.1",
-    "interweave": "^13.0.0",
+    "interweave": "^13.1.0",
     "jspdf": "^2.5.1",
     "lodash": "^4.17.21",
-    "node-html-parser": "^6.1.0",
-    "react": "^17.0.2",
+    "node-html-parser": "^6.1.5",
+    "react": "^18.2.0",
     "react-chartjs-2": "^5.2.0",
-    "react-dom": "^17.0.2",
-    "react-tabs": "^3.2.3",
-    "web-vitals": "^2.1.4"
+    "react-date-range": "^1.4.0",
+    "react-dom": "^18.2.0",
+    "react-tabs": "^6.0.2",
+    "web-vitals": "^3.4.0"
   },
   "scripts": {
     "start": "vite",
     "build": "vite build"
   },
   "devDependencies": {
-    "@vitejs/plugin-react": "^2.0.1",
-    "vite": "^3.0.7"
+    "@vitejs/plugin-react": "^4.0.4",
+    "vite": "^4.4.8"
   },
   "eslintConfig": {
     "extends": [
