@@ -1,8 +1,34 @@
+import { useState } from 'react';
 import Highcharts from 'highcharts';
 import HighchartsReact from 'highcharts-react-official';
+import {
+	Select,
+	MenuItem,
+	InputLabel,
+	FormControl,
+} from '@mui/material';
 import '../App.css';
 
-const getOptions = (type) => ({
+const selectColor = {
+	backgroundColor: 'white',
+	borderBottomLeftRadius: '4px',
+	borderBottomRightRadius: '4px',
+	outline: 'none',
+};
+
+const projectsType = ['confirmed', 'inprogress', 'default'];
+
+const data = (mapper) => {
+    const data =  {
+        confirmed: [25, 50, 2, 35, 15, 9],
+        inprogress: [10, 30, 37, 20, 39, 7],
+        default: [40, 19, 15, 50, 4, 24]
+    }
+
+    return data[mapper];
+}
+
+const getOptions = (type, mapper) => ({
     accessibility: {
         enabled: false
     },
@@ -17,7 +43,7 @@ const getOptions = (type) => ({
         height: 400,
     },
     title: {
-        text: 'Workflow Analytics',
+        text: `Workflow Analytics ${mapper.charAt(0).toUpperCase() + mapper.slice(1)}`,
     },
     xAxis: {
         categories: ['Smart Well', 'Single CHGP', 'Stack Selective', 'Stack Commingle', 'Open Hole HZ', 'Multizone']
@@ -30,7 +56,7 @@ const getOptions = (type) => ({
     series: [
         {
             name: 'Workflow Count',
-            data: [20, 5, 25, 45, 16, 9],
+            data: data(mapper) || [10, 20, 30, 40, 50, 45],
             color: 'rgb(25, 118, 210)'
         }
     ],
@@ -73,9 +99,34 @@ const donut = (type) => ({
 });
 
 function App() {
+    const [projectType, setProjectType] = useState('default');
+
+    console.log("edmvalue", projectType);
+
     return (
         <div className='highChart_div'>
-            <HighchartsReact highcharts={Highcharts} options={getOptions('bar')} />
+            <FormControl variant="filled" sx={{ m: 1, minWidth: 200, marginLeft: '30px', marginBottom: '40px' }}>
+                <InputLabel id="demo-simple-select-filled-label">
+                    Select Project Type
+                </InputLabel>
+                <Select
+                    labelId="demo-simple-select-filled-label"
+                    id="demo-simple-select-filled_edm-region-dropdown}"
+                    style={selectColor}
+                    label="Select Region"
+                    value={projectType}
+                    name="Region"
+                    onChange={(e) => {
+                        setProjectType(e.target.value);
+                    }}
+                >
+                    <MenuItem value="" key="first_Select_edm-region">
+                        Select Project Type
+                    </MenuItem>
+                    {projectsType.map(item => <MenuItem key={item} value={item}>{item}</MenuItem>)}
+                </Select>
+            </FormControl>
+            <HighchartsReact highcharts={Highcharts} options={getOptions('bar', projectType)} />
             <HighchartsReact highcharts={Highcharts} options={donut('pie')} />
         </div>
     );
